Moving User.php to Models/User.php & Updating config/auth.php
The User model is now in a dedicated Models folder which rather than the root.  This keeps the root free of clutter and makes the default laravel file structure more developer friendly right out of the box.  The User model namespace and Auth config have been updated appropriately.
